backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
alembic==1.13.1
asyncpg==0.29.0
httpx==0.25.2
python-dotenv==1.0.0
cryptography==41.0.7
pydantic-settings==2.1.0
websockets==12.0
bitcoinlib==0.12.2
web3==6.11.3
pylightning==0.1.1
aiofiles==23.2.1
structlog==23.2.0
slowapi==0.1.9
python-multipart==0.0.6
email-validator==2.1.0
jinja2==3.1.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
apscheduler==3.10.4
